servicios validation y processing ideam

+++ ADO.BL/DTOs/IdeamCompDTO.cs
@@ -3,9 +3,9 @@
     public class IdeamCompDTO
     {
         public string? code { get; set; }
-        public long? latitude { get; set; }
-        public long? longitude { get; set; }
-        public long? altitude { get; set; }
+        public float? latitude { get; set; }
+        public float? longitude { get; set; }
+        public float? altitude { get; set; }
         public string? department { get; set; }
         public string? municipality { get; set; }
     }
